feat(TodoList): adds tile rendering
feat(Tile): adds tile component
refactor(views): adds type file organization
